Totally updated and rebuild the JWT auth system. Removed deprecated code and start using fresh modules. Separated routes for each new module.

+++ internal/models/todo.go
@@ -10,7 +10,7 @@ type Todo struct {
 	Title       string
 	Description string
 	Completed   bool
-	UserID      uint // Foreign key for User (integer type)
+	UserID      float64
 	User        User // Associating with the User model
 	gorm.Model
 }
